please lint
git-svn-id: file:///svn/unbound/trunk@2008 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/tube.c
@@ -363,7 +363,7 @@ int tube_read_msg(struct tube* tube, uint8_t** buf, uint32_t* len,
 	}
 	d = 0;
 	while(d != (ssize_t)*len) {
-		if((r=read(fd, (*buf)+d, ((ssize_t)*len)-d)) == -1) {
+		if((r=read(fd, (*buf)+d, (size_t)((ssize_t)*len)-d)) == -1) {
 			log_err("tube msg read failed: %s", strerror(errno));
 			(void)fd_set_nonblock(fd);
 			free(*buf);
